Lots of updates, see full thing:
Changed options argument to arguments with defaults.

Made each element of the days array be an object with a day index and it's own startTime and endTime.
This is because some classes meet at different times on different days.
Simiarly, now if a class has nonstandard days, say MTuSu, it will correctly be one class object.